DELETEItem

+++ backend/controllers/DELETEItem.go
@@ -1,8 +1,6 @@
 package controllers
 
 import (
-	"strconv"
-
 	"gitee.com/masx200/to-do-list-go-sql-vue/backend/database"
 	"github.com/gin-gonic/gin"
 	"gorm.io/gorm"
@@ -10,32 +8,65 @@ import (
 
 func DELETEItem[T any](r *gin.Engine, createDB func() *gorm.DB, prefix string, model *T) {
 	r.DELETE(prefix, func(c *gin.Context) {
-		qsid := c.Query("id")
-		if len(qsid) == 0 {
-			c.String(400, "expect id but not found")
-			return
-		}
-		id, err := strconv.Atoi(qsid)
 
+		var inputs []map[string]any
+		err := c.ShouldBindJSON(&inputs)
 		if err != nil {
 			c.String(400, err.Error())
 			return
 		}
-		var item = new(T)
-		res, err := database.GetItem(createDB, item, uint(id))
-		/* 保持接口的幂等性 */
-		if err != nil {
-			c.JSON(200, []gin.H{{
-				"id": id,
-			}})
-			return
+		var ch = make(chan TWO[map[string]any, error])
+		for _, input := range inputs {
+			qsid, o := input["id"]
+			if !o {
+				c.String(400, "expect id but not found")
+				return
+			}
+			id, ok := qsid.(float64)
+
+			if !ok {
+				c.String(400, err.Error())
+				return
+			}
+
+			go func(id float64, ch chan TWO[map[string]any, error]) {
+
+				var item = new(T)
+				res, err := database.GetItem(createDB, item, uint(id))
+				/* 保持接口的幂等性 */
+				if err != nil {
+
+					ch <- TWO[map[string]any, error]{gin.H{
+						"id": id,
+					}, nil}
+
+					return
+				}
+				err = database.DeleteItem(createDB, new(T), uint(id))
+				if err != nil {
+					ch <- TWO[map[string]any, error]{nil, err}
+					return
+
+				} else {
+					ch <- TWO[map[string]any, error]{res, nil}
+					return
+
+				}
+			}(id, ch)
 		}
-		err = database.DeleteItem(createDB, new(T), uint(id))
-		if err != nil {
-			c.String(500, err.Error())
-		} else {
-			c.JSON(200, []map[string]interface{}{res})
+
+		var results = []map[string]interface{}{}
+		for range inputs {
+			two := <-ch
+			res := two.First
+			err := two.Second
+			if err != nil {
+				c.String(500, err.Error())
+				return
+			}
+			results = append(results, res)
 		}
-		// return
+		c.JSON(200, results)
+
 	})
 }
